rename fireBaseapi to api
