+++ src/Bom.js
@@ -3,7 +3,7 @@ import Product from "./component/Product";
 import Part from "./component/Part";
 import List from "./component/List";
 import Structure from "./component/Structure";
-import api from "./utils/firebaseApi";
+import api from "./utils/api";
 import * as S from "./component/StyleComponent";
 import styled from "styled-components";
