+++ src/Order.js
@@ -1,7 +1,7 @@
 import { Main } from "./component/StyleComponent";
 import Structure from "./component/Structure";
 import { useState, useEffect } from "react";
-import api from "./utils/firebaseApi";
+import api from "./utils/api";
 import styled from "styled-components";
 
 const Container = styled.div`
